refactor(web): extract Jira confirmation action bar
